New: make parser cache version- and interpreter-dependent New: basic support for PyPy (#426)
