+++ src/fandango/cli/__init__.py
@@ -1,7 +1,6 @@
 import argparse
 import atexit
 import glob
-import importlib.metadata
 import logging
 import os
 import os.path
@@ -57,23 +56,7 @@ from fandango.language.parse import parse
 from fandango.logger import LOGGER, print_exception
 
 from fandango import FandangoParseError, FandangoError
-
-
-DISTRIBUTION_NAME = "fandango-fuzzer"
-
-
-def version():
-    """Return the Fandango version number"""
-    return importlib.metadata.version(DISTRIBUTION_NAME)
-
-
-def homepage():
-    """Return the Fandango homepage"""
-    for key, value in importlib.metadata.metadata(DISTRIBUTION_NAME).items():
-        if key == "Project-URL" and value.startswith("homepage,"):
-            return value.split(",")[1].strip()
-    return "the Fandango homepage"
-
+import fandango
 
 def get_parser(in_command_line=True):
     # Main parser
@@ -92,7 +75,7 @@ def get_parser(in_command_line=True):
             Use `help COMMAND` to learn more about COMMAND.
             Use TAB to complete commands."""
         )
-    epilog += f"\nSee {homepage()} for more information."
+    epilog += f"\nSee {fandango.homepage()} for more information."
 
     main_parser = argparse.ArgumentParser(
         prog=prog,
@@ -106,7 +89,7 @@ def get_parser(in_command_line=True):
         main_parser.add_argument(
             "--version",
             action="version",
-            version=f"Fandango {version()}",
+            version=f"Fandango {fandango.version()}",
             help="show version number",
         )
 
@@ -1128,9 +1111,9 @@ def copyright_command(args):
 
 def version_command(args):
     if sys.stdout.isatty():
-        version_line = f"💃 {styles.color.ansi256(styles.rgbToAnsi256(128, 0, 0))}Fandango{styles.color.close} {version()}"
+        version_line = f"💃 {styles.color.ansi256(styles.rgbToAnsi256(128, 0, 0))}Fandango{styles.color.close} {fandango.version()}"
     else:
-        version_line = f"Fandango {version()}"
+        version_line = f"Fandango {fandango.version()}"
     print(version_line)
 
 
@@ -1439,14 +1422,6 @@ def main(*argv: str, stdout=sys.stdout, stderr=sys.stderr):
         last_status = 2
 
     return last_status
-
-
-def fandango(cmd: str, stdout=sys.stdout, stderr=sys.stderr):
-    # Entry point for tutorial
-    try:
-        main(*shlex.split(cmd, comments=True), stdout=stdout, stderr=stderr)
-    except SystemExit as e:
-        pass  # Do not exit
 
 
 if __name__ == "__main__":
